requirements.txt
requests==2.10.0
lxml==3.6.0
cssselect==0.9.1
python-dateutil==2.5.3
agate==1.4.0
